requirements.txt
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23